Updated OS detection in merc.h
I'm not sure how this used to work (did gcc in the past set a macro for the OS
name? I couldn't find anywhere else in the code that seemed like it would, nor
did any of the alternate Makefiles do that) but now it'll fall back to what was
previously the "#if defined(unix)" block if no other OS macro is defined.